% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/check_desc.R
\name{check_desc}
\alias{check_desc}
\title{Check a DESCRIPTION file on spelling errors.}
\usage{
check_desc(pkg_dir = getwd())
}
\arguments{
\item{pkg_dir}{a character vector of length one, specifying the location of a
package. The default value is the working directory (\code{\link{getwd}}).
Missing value will be ignored. The folder must contain file DESCRIPTION.}
}
\value{
A data.frame with columns: File, Line, Word. The File is DESCRIPTION
file; Line is the number of the line, where the error appears, and Word is
the word with error.
}
\description{
The function checks a package's DESCRIPTION file on spelling errors,
particularely Title: and Description: sections.
}
\details{
This function looks for the DESCRIPTION file in \code{pkg_dir} directory,
reads this file, and checks Title: and Discription: sections in typos with a
help of \code{hunspell} package.

\strong{Note:} Each line in section Description: must be intended by 4 spaces.
}
\seealso{
\code{\link{check_pkg}}, \code{\link{check_rd}}
}

