% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/check_rd.R
\name{check_rd}
\alias{check_rd}
\title{Check each .Rd file on spelling errors.}
\usage{
check_rd(pkg_dir = getwd(), sections = c("title", "desc", "details",
  "params", "value"))
}
\arguments{
\item{pkg_dir}{a character vector of length one, specifying the location of a
package. The default value is the working directory (\code{getwd()}). Missing
value will be ignored. The folder must contain a folder \code{/man} and
\code{/R}.}
}
\value{
A data.frame with columns: File, Line, Word. The File is .R file,
which generated correspondent .Rd file with an error, Line is the number of
the line, where the error appears, and Word is the word with error.
}
\description{
\code{check_rd} checks each .Rd generated file in folder \code{/man} on typos
and spelling errors.
}
\details{
This function goes through each file in folder \code{/man}, which has
extenssion .Rd and checks for a wrong spelling. Only limited set of sections
supported: title, description, details, parameters, and return. If the
section is missing, the function throws an massage. Custom sections are not
supported, due to limitations by package \code{Rd2roxygen}. The function uses
\code{hunspell} package to find typos. Instead of indicating the location of
the typo in .Rd file, the function seeks for the .R file in folder \code{/R},
which generated the .Rd file, and returns the line of the roxygen2 comment,
which generates such an error.
}
\seealso{
\code{\link{check_pkg()}}, \code{\link{check_desc()}}
}

